moved a few functions across files
